services/agent-runtime/app/models/base.py
"""Declarative base with ULID primary-key helper.

All models inherit from Base.  The generate_ulid() helper is used in default=
arguments so a new ULID is created at Python object construction time, before
the row is inserted.  This avoids a server-side RETURNING round-trip and makes
the ID available immediately after model instantiation.
"""

from datetime import datetime, timezone

from sqlalchemy import DateTime, func
from sqlalchemy.orm import DeclarativeBase, mapped_column, Mapped
from ulid import ULID


def generate_ulid() -> str:
    """Return a new ULID string (no prefix — callers add their own)."""
    return str(ULID())


class Base(DeclarativeBase):
    """Shared declarative base for all ORM models."""
    pass


class TimestampMixin:
    """Adds created_at / updated_at columns managed by the DB."""

    created_at: Mapped[datetime] = mapped_column(
        DateTime(timezone=True),
        server_default=func.now(),
        nullable=False,
    )
    updated_at: Mapped[datetime] = mapped_column(
        DateTime(timezone=True),
        server_default=func.now(),
        onupdate=func.now(),
        nullable=False,
    )
    deleted_at: Mapped[datetime | None] = mapped_column(
        DateTime(timezone=True),
        nullable=True,
        default=None,
    )

    def soft_delete(self) -> None:
        self.deleted_at = datetime.now(timezone.utc)

    @property
    def is_deleted(self) -> bool:
        return self.deleted_at is not None
